chore: add Edit Application functionality and refactor related components
